requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.0
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.3.1
urllib3==1.25.2
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.3.2
alembic==1.0.8
Flask_Bcrypt==0.7.1
passlib==1.7.2
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.3
python-dotenv==0.14.0
Flask_Migrate==2.5.2
requests==2.24.0
Werkzeug==0.14.1
